module-3/ml-admin-portfolio/requirements-dev.txt
anyio==4.2.0 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.10" and python_version < "4.0"
aws-cdk-asset-awscli-v1==2.2.201 ; python_version >= "3.10" and python_version < "4.0"
aws-cdk-asset-kubectl-v20==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
aws-cdk-asset-node-proxy-agent-v6==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
aws-cdk-aws-lambda-python-alpha==2.122.0a0 ; python_version >= "3.10" and python_version < "4.0"
aws-cdk-lib==2.122.0 ; python_version >= "3.10" and python_version < "4.0"
aws-lambda-powertools[aws-sdk]==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
awswrangler==3.5.1 ; python_version >= "3.10" and python_version < "4.0"
boto3-stubs[essential,sagemaker]==1.34.21 ; python_version >= "3.10" and python_version < "4.0"
boto3==1.34.21 ; python_version >= "3.10" and python_version < "4.0"
botocore-stubs==1.34.21 ; python_version >= "3.10" and python_version < "4.0"
botocore==1.34.21 ; python_version >= "3.10" and python_version < "4.0"
cattrs==23.2.3 ; python_version >= "3.10" and python_version < "4.0"
cdk-stacksets==0.0.149 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.10" and python_version < "4.0"
cfnresponse==1.1.2 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
cloudpickle==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
constructs==10.3.0 ; python_version >= "3.10" and python_version < "4.0"
contextlib2==21.6.0 ; python_version >= "3.10" and python_version < "4.0"
dataclasses-json==0.6.3 ; python_version >= "3.10" and python_version < "4.0"
dill==0.3.7 ; python_version >= "3.10" and python_version < "4.0"
docker==7.0.0 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
fastapi==0.95.2 ; python_version >= "3.10" and python_version < "4.0"
google-pasta==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.6 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==6.11.0 ; python_version >= "3.10" and python_version < "4.0"
importlib-resources==6.1.1 ; python_version >= "3.10" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
joblib==1.3.2 ; python_version >= "3.10" and python_version < "4.0"
jsii==1.94.0 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.21.0 ; python_version >= "3.10" and python_version < "4.0"
marshmallow==3.20.2 ; python_version >= "3.10" and python_version < "4.0"
multiprocess==0.70.15 ; python_version >= "3.10" and python_version < "4.0"
mypy-boto3-cloudformation==1.34.0 ; python_version >= "3.10" and python_version < "4.0"
mypy-boto3-dynamodb==1.34.21 ; python_version >= "3.10" and python_version < "4.0"
mypy-boto3-ec2==1.34.17 ; python_version >= "3.10" and python_version < "4.0"
mypy-boto3-lambda==1.34.0 ; python_version >= "3.10" and python_version < "4.0"
mypy-boto3-rds==1.34.6 ; python_version >= "3.10" and python_version < "4.0"
mypy-boto3-s3==1.34.14 ; python_version >= "3.10" and python_version < "4.0"
mypy-boto3-sagemaker==1.34.19 ; python_version >= "3.10" and python_version < "4.0"
mypy-boto3-sqs==1.34.0 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.26.3 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.1.4 ; python_version >= "3.10" and python_version < "4.0"
pathos==0.3.1 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==4.1.0 ; python_version >= "3.10" and python_version < "4.0"
pox==0.3.3 ; python_version >= "3.10" and python_version < "4.0"
ppft==1.7.6.7 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.25.2 ; python_version >= "3.10" and python_version < "4.0"
psutil==5.9.7 ; python_version >= "3.10" and python_version < "4.0"
publication==0.0.3 ; python_version >= "3.10" and python_version < "4.0"
pyarrow==14.0.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.13 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.10" and python_version < "4.0"
pywin32==306 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
referencing==0.32.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rpds-py==0.17.1 ; python_version >= "3.10" and python_version < "4.0"
s3transfer==0.10.0 ; python_version >= "3.10" and python_version < "4.0"
sagemaker==2.203.1 ; python_version >= "3.10" and python_version < "4.0"
schema==0.7.5 ; python_version >= "3.10" and python_version < "4.0"
scikit-learn==1.4.0 ; python_version >= "3.10" and python_version < "4.0"
scipy==1.11.4 ; python_version >= "3.10" and python_version < "4.0"
setuptools==69.0.3 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
smdebug-rulesconfig==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
tblib==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
threadpoolctl==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.66.1 ; python_version >= "3.10" and python_version < "4.0"
typeguard==2.13.3 ; python_version >= "3.10" and python_version < "4.0"
types-awscrt==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
types-s3transfer==0.10.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.10" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2023.4 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.18 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
xgboost==2.0.3 ; python_version >= "3.10" and python_version < "4.0"
yamldataclassconfig==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
zipp==3.17.0 ; python_version >= "3.10" and python_version < "4.0"
